+++ src/rpc/methods/eth.rs
@@ -306,8 +306,19 @@ pub enum Predefined {
     Latest,
 }
 
-#[derive(PartialEq, Debug, Clone, Serialize, Deserialize, Default, JsonSchema)]
-#[serde(rename_all = "camelCase")]
+#[derive(
+    PartialEq,
+    Debug,
+    Clone,
+    Serialize,
+    Deserialize,
+    Default,
+    JsonSchema,
+    strum::Display,
+    strum::EnumString,
+)]
+#[strum(serialize_all = "lowercase")]
+#[serde(rename_all = "lowercase")]
 pub enum ExtPredefined {
     Earliest,
     Pending,
@@ -413,6 +424,12 @@ pub enum ExtBlockNumberOrHash {
     BlockHashObject(BlockHash),
 }
 
+impl Default for ExtBlockNumberOrHash {
+    fn default() -> Self {
+        Self::PredefinedBlock(ExtPredefined::default())
+    }
+}
+
 lotus_json_with_self!(ExtBlockNumberOrHash);
 
 #[allow(dead_code)]
@@ -449,24 +466,13 @@ impl ExtBlockNumberOrHash {
     }
 
     pub fn from_str(s: &str) -> Result<Self, Error> {
-        match s {
-            "earliest" => Ok(ExtBlockNumberOrHash::from_predefined(
-                ExtPredefined::Earliest,
-            )),
-            "pending" => Ok(ExtBlockNumberOrHash::from_predefined(
-                ExtPredefined::Pending,
-            )),
-            "latest" | "" => Ok(ExtBlockNumberOrHash::from_predefined(ExtPredefined::Latest)),
-            "safe" => Ok(ExtBlockNumberOrHash::from_predefined(ExtPredefined::Safe)),
-            "finalized" => Ok(ExtBlockNumberOrHash::from_predefined(
-                ExtPredefined::Finalized,
-            )),
-            hex if hex.starts_with("0x") => {
-                let epoch = hex_str_to_epoch(hex)?;
-                Ok(ExtBlockNumberOrHash::from_block_number(epoch))
-            }
-            _ => Err(anyhow!("Invalid block identifier")),
+        if s.starts_with("0x") {
+            let epoch = hex_str_to_epoch(s)?;
+            return Ok(ExtBlockNumberOrHash::from_block_number(epoch));
         }
+        s.parse::<ExtPredefined>()
+            .map(ExtBlockNumberOrHash::from_predefined)
+            .map_err(|_| anyhow!("Invalid block identifier"))
     }
 }
 
@@ -4115,7 +4121,10 @@ fn get_eth_block_number_from_string<DB: Blockstore>(
     block: Option<&str>,
     resolve: ResolveNullTipset,
 ) -> Result<EthUint64> {
-    let block_param = ExtBlockNumberOrHash::from_str(block.unwrap_or("latest"))?;
+    let block_param = block
+        .map(ExtBlockNumberOrHash::from_str)
+        .transpose()?
+        .unwrap_or_default();
     Ok(EthUint64(
         tipset_by_ext_block_number_or_hash(chain_store, block_param, resolve)?.epoch() as u64,
     ))
@@ -4126,7 +4135,10 @@ async fn get_eth_block_number_from_string_v2<DB: Blockstore + Send + Sync + 'sta
     block: Option<&str>,
     resolve: ResolveNullTipset,
 ) -> Result<EthUint64> {
-    let block_param = ExtBlockNumberOrHash::from_str(block.unwrap_or("latest"))?;
+    let block_param = block
+        .map(ExtBlockNumberOrHash::from_str)
+        .transpose()?
+        .unwrap_or_default();
     Ok(EthUint64(
         tipset_by_block_number_or_hash_v2(ctx, block_param, resolve)
             .await?
@@ -4165,17 +4177,7 @@ impl RpcMethod<1> for EthTraceFilter {
         )
         .context("cannot parse toBlock")?;
 
-        Ok(trace_filter(ctx, filter, from_block, to_block)
-            .await?
-            .into_iter()
-            .sorted_by_key(|trace| {
-                (
-                    trace.block_number,
-                    trace.transaction_position,
-                    trace.trace.trace_address.clone(),
-                )
-            })
-            .collect_vec())
+        Ok(trace_filter(ctx, filter, from_block, to_block).await?)
     }
 }
 
@@ -4212,17 +4214,7 @@ impl RpcMethod<1> for EthTraceFilterV2 {
         .await
         .context("cannot parse toBlock")?;
 
-        Ok(trace_filter(ctx, filter, from_block, to_block)
-            .await?
-            .into_iter()
-            .sorted_by_key(|trace| {
-                (
-                    trace.block_number,
-                    trace.transaction_position,
-                    trace.trace.trace_address.clone(),
-                )
-            })
-            .collect::<Vec<_>>())
+        Ok(trace_filter(ctx, filter, from_block, to_block).await?)
     }
 }
 
@@ -4231,10 +4223,10 @@ async fn trace_filter(
     filter: EthTraceFilterCriteria,
     from_block: EthUint64,
     to_block: EthUint64,
-) -> Result<HashSet<EthBlockTrace>> {
+) -> Result<Vec<EthBlockTrace>> {
     let mut results = HashSet::default();
     if let Some(EthUint64(0)) = filter.count {
-        return Ok(results);
+        return Ok(Vec::new());
     }
     let count = *filter.count.unwrap_or_default();
     ensure!(
@@ -4245,7 +4237,8 @@ async fn trace_filter(
     );
 
     let mut trace_counter = 0;
-    for blk_num in from_block.0..=to_block.0 {
+    'blocks: for blk_num in from_block.0..=to_block.0 {
+        // For BlockNumber, EthTraceBlock and EthTraceBlockV2 are equivalent.
         let block_traces = EthTraceBlock::handle(
             ctx.clone(),
             (ExtBlockNumberOrHash::from_block_number(blk_num as i64),),
@@ -4266,7 +4259,7 @@ async fn trace_filter(
                 results.insert(block_trace);
 
                 if filter.count.is_some() && results.len() >= count as usize {
-                    return Ok(results);
+                    break 'blocks;
                 } else if results.len() > *FOREST_TRACE_FILTER_MAX_RESULT as usize {
                     bail!(
                         "too many results, maximum supported is {}, try paginating requests with After and Count",
@@ -4277,7 +4270,16 @@ async fn trace_filter(
         }
     }
 
-    Ok(results)
+    Ok(results
+        .into_iter()
+        .sorted_by_key(|trace| {
+            (
+                trace.block_number,
+                trace.transaction_position,
+                trace.trace.trace_address.clone(),
+            )
+        })
+        .collect_vec())
 }
 
 #[cfg(test)]