+++ CHANGELOG.md
@@ -33,6 +33,8 @@
 
 ### Changed
 
+- [#6522](https://github.com/ChainSafe/forest/pull/6522): `Filecoin.EthTraceFilter` filter options `from_block` and `to_block` now default to `latest` tag when omitted for v1 and v2 API.
+
 ### Removed
 
 ### Fixed